+++ src/rendering.ts
@@ -6,8 +6,10 @@ import type { Tile } from './tile';
 import type Zoom from './interaction';
 import type { AestheticSet } from './AestheticSet';
 import { timer, Timer } from 'd3-timer';
-import { Dataset, QuadtileDataset } from './Dataset';
-
+import { Dataset } from './Dataset';
+import type * as DS from './shared.d'
+import { Table } from 'apache-arrow';
+import { X } from './Aesthetic';
 class PlotSetting {
   start: number;
   value: number;
@@ -77,7 +79,7 @@ class RenderProps {
     this.foregroundSize = new PlotSetting(1, 'geometric');
     this.backgroundSize = new PlotSetting(1, 'geometric');
   }
-  apply_prefs(prefs: CompletePrefs) {
+  apply_prefs(prefs: DS.CompletePrefs) {
     const { duration } = prefs;
     this.maxPoints.update(prefs.max_points, duration);
     this.targetOpacity.update(prefs.alpha, duration);
@@ -126,9 +128,9 @@ export class Renderer<TileType extends Tile> {
   public height: number;
   public deferred_functions: Array<() => Promise<void> | void>;
   public _use_scale_to_download_tiles = true;
-  public zoom?: Zoom;
+  public zoom?: Zoom<TileType>;
   public aes?: AestheticSet<TileType>;
-  public _zoom?: Zoom;
+  public _zoom?: Zoom<TileType>;
   public _initializations: Promise<void>[] = [];
   public render_props: RenderProps = new RenderProps();
   constructor(
@@ -158,8 +160,8 @@ export class Renderer<TileType extends Tile> {
    * Render prefs are scatterplot prefs, but for a single tile
    * instead of for a whole table.
    */
-  get prefs(): RenderPrefs {
-    const p = { ...this.scatterplot.prefs } as RenderPrefs;
+  get prefs() {
+    const p = { ...this.scatterplot.prefs } as DS.CompletePrefs & {arrow_table?: Table, arrow_buffer?: Uint8Array };
     // Delete the arrow stuff b/c serializing it is crazy expensive.
     p.arrow_table = undefined;
     p.arrow_buffer = undefined;
@@ -184,8 +186,8 @@ export class Renderer<TileType extends Tile> {
     const pixel_area = (width * height) / pixelRatio;
     const total_intended_points = min([
       max_ix,
-      (this.dataset.highest_known_ix as number | undefined) || 1e10,
-    ]) as number;
+      (this.dataset.highest_known_ix) || 1e10,
+    ]);
 
     const total_points = total_intended_points * (1 - discard_share);
 
@@ -222,6 +224,8 @@ export class Renderer<TileType extends Tile> {
     const { max_ix } = this;
     const { dataset: tileSet } = this;
     // Materialize using a tileset method.
+
+    const x = this.aes.dim('x') as X;
     const natural_display =
       this.aes.dim('x').current.field == 'x' &&
       this.aes.dim('y').current.field == 'y' &&